Use impl_ops instead of hand-written macros
