+++ plutus-ledger-api/Cargo.toml
@@ -21,6 +21,7 @@ data-encoding = { version = "^2.4.0", optional = true }
 thiserror = "^1.0.50"
 linked-hash-map = "~0.5.6"
 num-traits = "~0.2.17"
+impl_ops = "0.1.1"
 
 [features]
 serde = ["dep:serde", "num-bigint/serde"]
