--- plutus-ledger-api/src/utils/value/mod.rs
@@ -1,3 +1,0 @@
-#[macro_use]
-pub mod macros;
-pub mod value_utils;
